refact: move PicoSvg into pico_svg module
